requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==1.3.5  # align with app specific pin (was >=1.0.0)
httpx==0.24.1
python-dotenv==1.0.1  # align with app pin (was 0.21.0 minimum)
requests==2.31.0
gunicorn==20.1.0
pytz==2024.1
gevent==24.2.1

# RAG & data processing dependencies
pandas>=1.3.0
openpyxl>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
python-docx>=0.8.11
PyPDF2>=1.26.0
scikit-learn>=0.24.0

# MCP wrapper runtime (implicit via requests) - keep explicit for clarity
# (No extra lines needed; requests already included.)

# Testing
pytest>=8.0.0

